requirements.txt
langchain-core>=0.1.30
langchain-anthropic>=0.1.4
langchain-openai>=0.0.8
langchain-google-genai>=0.0.11
python-dotenv>=1.0.1
langgraph>=0.0.26 